Minor and documentation changes
DDICODE-533
Improve documentation of attributes var and varGrp attributes in varGrp
The declaration of the attribute type is located only in the hierarchy and is not repeated in the documentation. 
CHANGED documentation in varGrp element to clarify use of attributes var and varGrp. Also updated parallel documentation in nCubeGrp element.

DDICODE-87
New attribute titl for grantNo element
ADDED fundingProgram and grantName to grantNo and updated documentation
PROBLEM see issue

DDICODE-70
Geographic Bounding Box although 0..1 in Summary Description can be replicated by use of multiple sumDscr
ADDED documentation to geoBndBox and to sumDscr and boundPoly
Replication of the element geoBndBox is NOT recommended. The purpose of a bounding box is to support high level geographic point search systems. Most search systems of this type do not handle multiple instances of a bounding box. The bounding box should represent the full geographic coverage extent of the of the overall datasets being described by the Codebook instance. If there is a desire to provide the equivilent of a bounding box for each of multiple summary descriptions, use of the boundPoly is recommended. First provide a geoBndBox for the full area covered by the study. Then provide a boundPoly for each geographic area defined with a separate sumDscr. Note that when describing a bounding box using a boundPoly description the four corner points are described. The starting point and end point should match (closing the polygon).   
Clarified language, usage, and examples

DDICODE-100
possibility to document interviewee/participant instructions