Add files via upload
ISSUE #15
If the attribute "agentIdentifier" is used, "typeOfAgentIdentifier" should also be provided. An agent may be an organization or individual. Use "isPersistantIdentifier" (true|false) to indicate if this is intended to be a persistent identifier. Use "agentType" to indicate if the agent is an organization or an individual.

Corrected examples to reflect revised attribute names agentIdentifier and typeOfAgentIdentifier
contact
distrbtr
metadataAccs
origArch
othID

ISSUE #17 check...add standard documenation

Corrected misspelling of typeOfAccess in documentation text
dataAccs
metadataAccs

Updated the documentation of the following
typeOfAccess,
digitalFingerprintValue, Provide the digital fingerprint in digitalFingerprintValue and identify the algorithm specification used (add version as a separate entry if it is not part of the specification entry). [changed to add proper element/attribute name in quotations]
algorithmSpecification, [see digitalFingerprintValue]
algorithmVersion, [see digitalFingerprintValue]
description, added "" as needed
outcome, added "" as needed
otherQualityStatement, added "" as needed
complianceDescription, added "" as needed

ADDED elements and documentation for:
typeOfExPostEvaluation
typeOfCodingInstruction,
typeOfOtherMaterial,
typeOfSetAvailability,
typeOfDataSrc,
typeOfDevelopmentActivity,
               
ISSUE #19
In software element CHAGE Software used to produce the work. Supprts  TO Software used to produce the work. Supports 

ISSUE #21
Added documenation to catGrp
Use the attribute "catgry" (IDREFS type) to provide a space delimited list of the ID of each category in the parent category group. The attribute "catGrp" (IDREFS type) is used to indicate all the subsidiary category groups which nest underneath the current category group. This allows for the encoding of a hierarchical structure of category groups. Do not list the categories contained in a listed "catGrp" within the attribute "catgry". These two IDREFS attributes should list ONLY the direct children of the parent category group.

ISSUE #23
11 instances
CHANGED documenation FROM "an "vocabAgencyName" and "vocabAgencyID" to specify the agency managing" TO "a "vocabAgencyName" to specify the agency managing" in all documentation

ISSUE #24
Revised documentation to: 
The focus of this element is on the original archive, differentiating between the designated depository and any locally held copies obtained to support local use.

ISSUE #25
REMOVED as redundant "An agent may be an organization or individual." holdover from when there was no indictor for individual or ogranization status.

ISSUE #26
CHANGED documentation on all conceptType or extension base conceptType FROM "The string content of "concept"..." TO "The string content of the element..."

ISSUE #27
typeOfAccs now has documentation on use and this documentation has been removed from parents metadataAccs and dataAccs