Add files via upload
Issue #14
Improve documentation for Question adding a note about attribute ID as this is one place where the use of an ID is needed to support common usage of IDREF or IDREFS
Also corrected the spelling of sub-element in setAvail

Issue #29
Reword documentation regarding CV support fileType
Changed documenation to: The use a standard term from a controlled vocabulary for the attribute "charset" is recommended even though the attribute structure does not allow for the declaration of the controlled vocabulary used. DDI provides a Controlled Vocabulary for this location: "CharacterSet"

Issue #30
correct all misspelling of ORCID or orcid as ORCHID/orchid
Checked entire xsd. found and corrected one instance of misspelling in "contact" example